Change Value type from u8 to i64 for improved range; update main.rs for clarity and additional operations

+++ src/lib.rs
@@ -1,4 +1,5 @@
-pub type Value = u8;
+//pub type Value = u8;
+pub type Value = i64;
 
 //operation codes
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
